It uses the current font size instead of fl_height() to measure text vertically.  This I think it the source of differences in appearance between X and Win32.  On X the font reports a height that is often much larger than the point size used to select it.  This also requires widgets to open the display just to measure themselves.  With this fix the default "leading" of 2 looks correct (and in fact it cannot be less than 2 for some X fonts, although True Type fonts are ok).
While I was at it I changed fl_width to return int instead of double.
Almost every call just cast it to int anyway.  (if a system ever
returns float we can add a new call for that)

Fl_Tile and Fl_Tabs do their callback() on changes made by the user.


git-svn-id: file:///fltk/svn/trunk@878 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ FL/gl.h
@@ -1,5 +1,5 @@
 //
-// "$Id: gl.h,v 1.9 1999/10/17 20:40:15 vincent Exp $"
+// "$Id: gl.h,v 1.10 1999/11/16 07:36:05 bill Exp $"
 //
 // OpenGL header file for the Fast Light Tool Kit (FLTK).
 //
@@ -54,9 +54,9 @@ inline void gl_rectf(int x,int y,int w,int h) {glRecti(x,y,x+w,y+h);}
 void gl_font(Fl_Font f, int size);
 int  gl_height();
 int  gl_descent();
-double gl_width(const char *);
-double gl_width(const char *, int n);
-double gl_width(uchar);
+int  gl_width(const char *);
+int  gl_width(const char *, int n);
+int  gl_width(uchar);
 
 void gl_draw(const char*);
 void gl_draw(const char*, int n);
@@ -72,5 +72,5 @@ void gl_draw_image(const uchar *, int x,int y,int w,int h, int d=3, int ld=0);
 #endif
 
 //
-// End of "$Id: gl.h,v 1.9 1999/10/17 20:40:15 vincent Exp $".
+// End of "$Id: gl.h,v 1.10 1999/11/16 07:36:05 bill Exp $".
 //
